black and lint
