+++ backend/api/models/data/dataset.py
@@ -5,15 +5,45 @@ from pathlib import Path
 
 from django.conf import settings
 from django.db import models
+from django.db.models import Manager
 from metadatax.acquisition.models import ChannelConfiguration
 from osekit.public_api.dataset import Dataset as OSEkitDataset
 from typing_extensions import deprecated
 
+from backend.aplose.models import User
 from .__abstract_dataset import AbstractDataset
+
+
+class DatasetManager(Manager):  # pylint: disable=too-few-public-methods
+    """Dataset manager"""
+
+    def get_or_create(self, name: str, path: str, owner: User, legacy: bool = False):
+        """Get or create dataset, use owner only for creation"""
+        if Dataset.objects.filter(name=name, path=path).exists():
+            return (
+                Dataset.objects.get(
+                    name=name,
+                    path=path,
+                    legacy=legacy,
+                ),
+                False,
+            )
+
+        return (
+            Dataset.objects.create(
+                name=name,
+                path=path,
+                owner=owner,
+                legacy=legacy,
+            ),
+            True,
+        )
 
 
 class Dataset(AbstractDataset, models.Model):
     """Dataset"""
+
+    objects = DatasetManager()
 
     class Meta:
         unique_together = (
