+++ backend/api/schema/data/__init__.py
@@ -2,13 +2,14 @@
 
 import graphene
 
-## Query
 from .colormap__query import ColormapQuery
+from .dataset__mutation import DatasetMutation
 from .dataset__query import DatasetQuery
 from .fft__query import FFTQuery
 from .legacy_spectrogram_configuration__query import LegacySpectrogramConfigurationQuery
 from .scales__query import ScalesQuery
 from .spectrogram__query import SpectrogramQuery
+from .spectrogram_analysis__mutation import SpectrogramAnalysisMutation
 from .spectrogram_analysis__query import SpectrogramAnalysisQuery
 
 
@@ -23,11 +24,6 @@ class APIDataQuery(
     graphene.ObjectType,
 ):  # pylint: disable=too-few-public-methods
     """API GraphQL queries"""
-
-
-## Mutation
-from .dataset__mutation import DatasetMutation
-from .spectrogram_analysis__mutation import SpectrogramAnalysisMutation
 
 
 class APIDataMutation(
